// src/app/admin/pages/experience/experience.types.ts
// Copyright 2025 josephmbote
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     https://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

export interface ExperienceItem {
  id: string;
  position: string;
  company: string;
  period: string;
  description: string;
  achievements: string[];
}

export type ExperienceDraft = Omit<ExperienceItem, 'id'>;